requirements.txt
Flask>=3.0
hashlib
sys
pycryptodome
